requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.7
Pillow==10.0.0
Flask-CKEditor==0.4.6
slugify==0.0.1 # Added for slug generation
python-slugify==8.0.4 # Dependency for slugify
Flask-Bootstrap==3.3.7.1 # Added Flask-Bootstrap
